+++ src/commands/query_json.ml
@@ -226,6 +226,10 @@ let dump (type a) : a t -> json =
                                          | `Unqualify -> "unqualify");
       "position", mk_position pos;
     ]
+  | Signature_help pos ->
+    mk "signature-help" [
+      "position", mk_position pos
+    ]
   | Version -> mk "version" []
 
 let string_of_completion_kind = function
@@ -367,6 +371,22 @@ let json_of_inlay_hints hints =
      ]
   in `List (List.map ~f:json_of_hint hints)
 
+let json_of_signature_help resp =
+  let param { label_start; label_end } =
+    `Assoc ["label", `List [`Int label_start; `Int label_end]] in
+  match resp with
+  | None -> `Assoc []
+  | Some { label; parameters; active_param; active_signature } ->
+    let signature =
+      `Assoc
+        ["label", `String label;
+          "parameters", `List (List.map ~f:param parameters);] in
+    `Assoc
+      ["signatures", `List [signature];
+       "activeParameter", `Int active_param;
+       "activeSignature", `Int active_signature;
+      ]
+
 let json_of_response (type a) (query : a t) (response : a) : json =
   match query, response with
   | Type_expr _, str -> `String str
@@ -400,7 +420,7 @@ let json_of_response (type a) (query : a t) (response : a) : json =
       | `Found doc ->
         `String doc
     end
-  | Syntax_document _, resp -> 
+  | Syntax_document _, resp ->
     (match resp with
     | `Found info ->
       `Assoc
@@ -410,9 +430,9 @@ let json_of_response (type a) (query : a t) (response : a) : json =
         ("url", `String info.documentation);
       ]
     | `No_documentation -> `String "No documentation found")
-  | Expand_ppx _, resp -> 
+  | Expand_ppx _, resp ->
     let str = match resp with
-    | `Found ppx_info -> 
+    | `Found ppx_info ->
       `Assoc
       [
         ("code", `String ppx_info.code);
@@ -471,5 +491,6 @@ let json_of_response (type a) (query : a t) (response : a) : json =
     let with_file = scope = `Project in
     `List (List.map locations
              ~f:(fun loc -> with_location ~with_file loc []))
+  | Signature_help _, s -> json_of_signature_help s
   | Version, version ->
     `String version